requirements.txt
# Core Data Pipeline Dependencies
telethon           # Telegram scraping
sqlalchemy         # Database interaction
python-dotenv      # Environment variable management
pydantic           # Data validation (used by FastAPI)
psycopg2-binary    # PostgreSQL connectivity for loader script

# Data Transformation
dbt-postgres       # Data transformation and modeling

# API
fastapi            # Analytical API
uvicorn            # ASGI server for FastAPI

# Orchestration
dagster
dagster-webserver

# Data Enrichment
ultralytics        # YOLOv8 for object detection

# Utilities
tqdm
requests           # For API usage examples and scripts
jupyter            # For running notebooks (EDA, scraping)

# Testing
pytest 
httpx            # For async HTTP requests in tests 